feat: organize scripts better and improve file search
