+++ package.json
@@ -31,7 +31,7 @@
     "check-ts-references": "pnpm update-ts-references --check",
     "knip": "knip",
     "postinstall": "pnpm update-ts-references",
-    "add-tailwind-darkmode-classes": "tsx scripts/add-tailwind-darkmode-classes.ts"
+    "add-dark-mode": "tsx scripts/dark-mode/main.ts"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.8.1",
