+++ pnpm-workspace.yaml
@@ -9,6 +9,7 @@ packages:
   - "!releases"
 
 catalog:
+  "execa": 9.6.1
   "vite-tsconfig-paths": 6.1.1
   "@modelcontextprotocol/sdk": 1.29.0
   "@powerhousedao/document-engineering": 1.40.5
