--- scripts/dark-mode-candidate-files.ts
@@ -1,228 +1,0 @@
-export const darkModeCandidateFiles = [
-  "./apps/connect/src/components/app-skeleton.tsx",
-  "./apps/connect/src/components/button.tsx",
-  "./apps/connect/src/components/cookie-banner.tsx",
-  "./apps/connect/src/components/editors.tsx",
-  "./apps/connect/src/components/error-boundary.tsx",
-  "./apps/connect/src/components/migration-banner.tsx",
-  "./apps/connect/src/components/missing-model-banner.tsx",
-  "./apps/connect/src/components/modal/modals/CookiesPolicyModal.tsx",
-  "./apps/connect/src/components/modal/modals/DebugSettingsModal.tsx",
-  "./apps/connect/src/components/modal/modals/DownloadDocumentWithErrorsModal.tsx",
-  "./apps/connect/src/components/modal/modals/SettingsModal.tsx",
-  "./apps/connect/src/components/modal/modals/settings/about.tsx",
-  "./apps/connect/src/components/search-bar.tsx",
-  "./apps/connect/src/pages/settings.tsx",
-  "./packages/codegen/src/templates/app/components/CreateDocument.ts",
-  "./packages/codegen/src/templates/app/components/EmptyState.ts",
-  "./packages/codegen/src/templates/app/components/Files.ts",
-  "./packages/codegen/src/templates/app/components/Folders.ts",
-  "./packages/codegen/src/templates/app/components/NavigationBreadcrumbs.ts",
-  "./packages/codegen/src/templates/app/editor.ts",
-  "./packages/codegen/src/templates/boilerplate/LICENSE.ts",
-  "./packages/codegen/src/templates/document-editor/editor.ts",
-  "./packages/design-system/src/connect/components/account-popover/account-popover-login.tsx",
-  "./packages/design-system/src/connect/components/account-popover/account-popover-user.tsx",
-  "./packages/design-system/src/connect/components/breadcrumbs/breadcrumbs.tsx",
-  "./packages/design-system/src/connect/components/code-popover.tsx",
-  "./packages/design-system/src/connect/components/combobox/combobox.tsx",
-  "./packages/design-system/src/connect/components/cookie-banner/cookie-banner.tsx",
-  "./packages/design-system/src/connect/components/db-explorer/components/filter-bar.tsx",
-  "./packages/design-system/src/connect/components/db-explorer/components/schema-tree-sidebar.tsx",
-  "./packages/design-system/src/connect/components/db-explorer/components/table-view.tsx",
-  "./packages/design-system/src/connect/components/db-explorer/db-explorer.tsx",
-  "./packages/design-system/src/connect/components/debug-inspector/debug-inspector.tsx",
-  "./packages/design-system/src/connect/components/default-editor-loader/default-editor-loader.tsx",
-  "./packages/design-system/src/connect/components/disclosure/disclosure.tsx",
-  "./packages/design-system/src/connect/components/divider/divider.tsx",
-  "./packages/design-system/src/connect/components/document-state-viewer/document-state-viewer.tsx",
-  "./packages/design-system/src/connect/components/document-timeline/components/h-divider.tsx",
-  "./packages/design-system/src/connect/components/document-timeline/components/timeline-bar.tsx",
-  "./packages/design-system/src/connect/components/document-timeline/document-timeline.tsx",
-  "./packages/design-system/src/connect/components/document-toolbar/containers.tsx",
-  "./packages/design-system/src/connect/components/document-toolbar/toolbar-button.tsx",
-  "./packages/design-system/src/connect/components/document-toolbar/toolbar-input.tsx",
-  "./packages/design-system/src/connect/components/document-toolbar/toolbar-name.tsx",
-  "./packages/design-system/src/connect/components/drop-zone/drop-zone.tsx",
-  "./packages/design-system/src/connect/components/dropdown-menu/dropdown-menu.tsx",
-  "./packages/design-system/src/connect/components/editor-action-buttons/editor-action-buttons.tsx",
-  "./packages/design-system/src/connect/components/editor-undo-redo-buttons/editor-undo-redo-buttons.tsx",
-  "./packages/design-system/src/connect/components/file-item/file-item.tsx",
-  "./packages/design-system/src/connect/components/folder-item/folder-item.tsx",
-  "./packages/design-system/src/connect/components/footer/footer.tsx",
-  "./packages/design-system/src/connect/components/form-input/form-input.tsx",
-  "./packages/design-system/src/connect/components/form/add-local-drive-form.tsx",
-  "./packages/design-system/src/connect/components/form/add-remote-drive-form.tsx",
-  "./packages/design-system/src/connect/components/form/drive-settings-form.tsx",
-  "./packages/design-system/src/connect/components/form/inputs/available-offline-toggle.tsx",
-  "./packages/design-system/src/connect/components/form/inputs/delete-drive.tsx",
-  "./packages/design-system/src/connect/components/form/inputs/drive-app.tsx",
-  "./packages/design-system/src/connect/components/form/inputs/drive-name.tsx",
-  "./packages/design-system/src/connect/components/form/inputs/label.tsx",
-  "./packages/design-system/src/connect/components/form/inputs/location-info.tsx",
-  "./packages/design-system/src/connect/components/formatted-json-viewer.tsx",
-  "./packages/design-system/src/connect/components/home-screen/home-screen-item.tsx",
-  "./packages/design-system/src/connect/components/home-screen/home-screen.tsx",
-  "./packages/design-system/src/connect/components/image-input/image-input.tsx",
-  "./packages/design-system/src/connect/components/integrity-inspector/integrity-inspector.tsx",
-  "./packages/design-system/src/connect/components/loading-screen/loading-screen.tsx",
-  "./packages/design-system/src/connect/components/modal/add-local-drive-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/add-remote-drive-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/confirmation-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/create-document-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/delete-drive-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/delete-item-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/drive-settings-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/inspector-modal/inspector-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/missing-package-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/read-required-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/replace-duplicate-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal-v2/about.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal-v2/danger-zone.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal-v2/default-editor.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal-v2/package-manager/package-manager-input.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal-v2/package-manager/package-manager-list.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal-v2/package-manager/version-picker.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal-v2/settings-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal/clear-storage-row.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal/dependency-versions/dependency-versions.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal/row.tsx",
-  "./packages/design-system/src/connect/components/modal/settings-modal/settings-modal.tsx",
-  "./packages/design-system/src/connect/components/modal/upgrade-drive-modal.tsx",
-  "./packages/design-system/src/connect/components/node-input/node-input.tsx",
-  "./packages/design-system/src/connect/components/object-inspector-modal/object-inspector-modal.tsx",
-  "./packages/design-system/src/connect/components/processors-inspector/processors-inspector.tsx",
-  "./packages/design-system/src/connect/components/queue-inspector/queue-inspector.tsx",
-  "./packages/design-system/src/connect/components/remotes-inspector/components/channel-inspector.tsx",
-  "./packages/design-system/src/connect/components/remotes-inspector/components/connection-state-badge.tsx",
-  "./packages/design-system/src/connect/components/remotes-inspector/components/mailbox-table.tsx",
-  "./packages/design-system/src/connect/components/remotes-inspector/remotes-inspector.tsx",
-  "./packages/design-system/src/connect/components/revision-history/header/branch.tsx",
-  "./packages/design-system/src/connect/components/revision-history/header/doc-id.tsx",
-  "./packages/design-system/src/connect/components/revision-history/header/document-state.tsx",
-  "./packages/design-system/src/connect/components/revision-history/header/header.tsx",
-  "./packages/design-system/src/connect/components/revision-history/header/scope.tsx",
-  "./packages/design-system/src/connect/components/revision-history/revision-history.tsx",
-  "./packages/design-system/src/connect/components/revision-history/revision/address.tsx",
-  "./packages/design-system/src/connect/components/revision-history/revision/errors.tsx",
-  "./packages/design-system/src/connect/components/revision-history/revision/operation.tsx",
-  "./packages/design-system/src/connect/components/revision-history/revision/revision-number.tsx",
-  "./packages/design-system/src/connect/components/revision-history/revision/revision.tsx",
-  "./packages/design-system/src/connect/components/revision-history/revision/signature.tsx",
-  "./packages/design-system/src/connect/components/revision-history/revision/timestamp.tsx",
-  "./packages/design-system/src/connect/components/revision-history/skip/skip.tsx",
-  "./packages/design-system/src/connect/components/revision-history/timeline/day.tsx",
-  "./packages/design-system/src/connect/components/revision-history/timeline/timeline.tsx",
-  "./packages/design-system/src/connect/components/search-bar/filter-item.tsx",
-  "./packages/design-system/src/connect/components/search-bar/search-bar.tsx",
-  "./packages/design-system/src/connect/components/select/select.tsx",
-  "./packages/design-system/src/connect/components/sidebar/sidebar-footer.tsx",
-  "./packages/design-system/src/connect/components/sidebar/sidebar-header.tsx",
-  "./packages/design-system/src/connect/components/sidebar/sidebar-item.tsx",
-  "./packages/design-system/src/connect/components/sidebar/sidebar-login.tsx",
-  "./packages/design-system/src/connect/components/status-icon/sync-status-icon.tsx",
-  "./packages/design-system/src/connect/components/tabs/tabs.tsx",
-  "./packages/design-system/src/connect/components/toast/toast.tsx",
-  "./packages/design-system/src/connect/components/toggle/toggle.tsx",
-  "./packages/design-system/src/connect/components/tooltip/tooltip.tsx",
-  "./packages/design-system/src/connect/components/upload-file-item/components/error-details.tsx",
-  "./packages/design-system/src/connect/components/upload-file-item/components/header.tsx",
-  "./packages/design-system/src/connect/components/upload-file-item/components/progress-bar.tsx",
-  "./packages/design-system/src/connect/components/upload-file-item/components/status-row.tsx",
-  "./packages/design-system/src/connect/components/upload-file-item/upload-file-item.tsx",
-  "./packages/design-system/src/connect/components/upload-file-list/upload-file-list.tsx",
-  "./packages/design-system/src/connect/constants/options.tsx",
-  "./packages/design-system/src/powerhouse/components/button/button.tsx",
-  "./packages/design-system/src/powerhouse/components/legacy/styles.css",
-  "./packages/design-system/src/powerhouse/components/legacy/styles.ts",
-  "./packages/design-system/src/powerhouse/components/legacy/stylesVariant.ts",
-  "./packages/design-system/src/powerhouse/components/modal/modal.tsx",
-  "./packages/design-system/src/powerhouse/components/pagination/pagination-button.tsx",
-  "./packages/design-system/src/powerhouse/components/sidebar/sidebar-panel.tsx",
-  "./packages/design-system/src/powerhouse/components/sidebar/sidebar.tsx",
-  "./packages/design-system/src/ui/components/button/button.tsx",
-  "./packages/design-system/src/ui/components/character-counter/character-counter.tsx",
-  "./packages/design-system/src/ui/components/checkbox/checkbox-base.tsx",
-  "./packages/design-system/src/ui/components/command/command.tsx",
-  "./packages/design-system/src/ui/components/decorators.tsx",
-  "./packages/design-system/src/ui/components/form-description/form-description.tsx",
-  "./packages/design-system/src/ui/components/form-label/form-label.tsx",
-  "./packages/design-system/src/ui/components/form-message/form-message.tsx",
-  "./packages/design-system/src/ui/components/form-message/message-list.tsx",
-  "./packages/design-system/src/ui/components/id-autocomplete/id-autocomplete-input-container.tsx",
-  "./packages/design-system/src/ui/components/id-autocomplete/id-autocomplete-list-option.tsx",
-  "./packages/design-system/src/ui/components/id-autocomplete/id-autocomplete-list.tsx",
-  "./packages/design-system/src/ui/components/id-autocomplete/id-autocomplete.tsx",
-  "./packages/design-system/src/ui/components/index.ts",
-  "./packages/design-system/src/ui/components/input/input.tsx",
-  "./packages/design-system/src/ui/components/input/splitted-input-diff.tsx",
-  "./packages/design-system/src/ui/components/input/subcomponent/input-diff.tsx",
-  "./packages/design-system/src/ui/components/input/subcomponent/text-diff.tsx",
-  "./packages/design-system/src/ui/components/json-viewer/json-viewer.tsx",
-  "./packages/design-system/src/ui/components/popover/popover.tsx",
-  "./packages/design-system/src/ui/components/radio-group-field/radio.tsx",
-  "./packages/design-system/src/ui/components/search-autocomplete/search-autocomplete.tsx",
-  "./packages/design-system/src/ui/components/select-field/content.tsx",
-  "./packages/design-system/src/ui/components/select-field/select-field.tsx",
-  "./packages/design-system/src/ui/components/select-field/selected-content.tsx",
-  "./packages/design-system/src/ui/components/select-field/subcomponents/CommandItemList.tsx",
-  "./packages/design-system/src/ui/components/text-field/text-field.tsx",
-  "./packages/design-system/src/ui/components/text-input/text-input.tsx",
-  "./packages/design-system/src/ui/components/tooltip/tooltip.tsx",
-  "./packages/design-system/style.css",
-  "./packages/design-system/theme.css",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/button.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/divider.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/errors.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/form.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/input.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/label.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/model-metadata-form.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/module-form.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/module.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/operation-description-form.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/operation-error-form.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/operation-form.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/operation.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/state-schemas.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/tabs.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/text-area.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/components/text-field.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/editor.tsx",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/utils/helpers.test.ts",
-  "./packages/powerhouse-vetra-packages/editors/document-model-editor/utils/helpers.ts",
-  "./packages/powerhouse-vetra-packages/editors/generic-drive-explorer/components/create-document.tsx",
-  "./packages/powerhouse-vetra-packages/editors/generic-drive-explorer/components/file-content-view.tsx",
-  "./packages/powerhouse-vetra-packages/editors/generic-drive-explorer/components/folder-view.tsx",
-  "./packages/powerhouse-vetra-packages/editors/generic-drive-explorer/components/layout.tsx",
-  "./packages/powerhouse-vetra-packages/editors/generic-drive-explorer/components/search-bar.tsx",
-  "./packages/reactor-browser/src/renown/components/RenownLoginButton.tsx",
-  "./packages/reactor-browser/src/renown/components/RenownUserButton.tsx",
-  "./packages/reactor-browser/src/utils/index.ts",
-  "./packages/vetra/editors/app-editor/components/AppEditorForm.tsx",
-  "./packages/vetra/editors/app-editor/components/EditName.tsx",
-  "./packages/vetra/editors/app-editor/editor.tsx",
-  "./packages/vetra/editors/components/StatusPill.tsx",
-  "./packages/vetra/editors/document-editor/components/DocumentEditorForm.tsx",
-  "./packages/vetra/editors/document-editor/editor.tsx",
-  "./packages/vetra/editors/processor-editor/components/ProcessorEditorForm.tsx",
-  "./packages/vetra/editors/processor-editor/editor.tsx",
-  "./packages/vetra/editors/subgraph-editor/components/SubgraphEditorForm.tsx",
-  "./packages/vetra/editors/subgraph-editor/editor.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/Accordion.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/DataIntegrationsColumn.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/DocumentModelsColumn.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/DriveHeader.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/DriveInfoItem.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/ModuleAccordion.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/ModuleItem.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/ModuleList.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/NewModuleItem.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/PackageInformationSection.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/SectionAccordion.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/ShareMenuItem.tsx",
-  "./packages/vetra/editors/vetra-drive-app/components/UserExperiencesColumn.tsx",
-  "./packages/vetra/editors/vetra-drive-app/editor.tsx",
-  "./packages/vetra/editors/vetra-package/components/MetaForm.tsx",
-  "./packages/vetra/editors/vetra-package/editor.tsx",
-];
