+++ scripts/dark-mode/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../tsconfig.options.json",
+  "compilerOptions": {
+    "types": ["node"],
+    "resolveJsonModule": true,
+    "noEmit": true,
+    "moduleResolution": "NodeNext",
+    "skipLibCheck": true
+  },
+  "include": ["./*.ts"],
+  "exclude": ["node_modules", "dist"]
+}
